Cleanup of SolarSystem
Texture paths are fetched via helper, createPlanet was separated, generateDescription was separated,  many variables were renamed and re-ordered.
