+++ components/Panorama/Panorama.js
