+++ pages/index.js
@@ -3,7 +3,7 @@ import Image from 'next/image'
 import Link from "next/link"
 import styles from '../styles/Home.module.css'
 
-import Panorama from "../components/Panorama"
+import Panorama from "../components/Panorama/Panorama"
 
 /**
  * Creates the Home page
